Refactor components to remove no-multi-comp suppressions
