+++ components/map/MapControls.tsx
@@ -5,9 +5,7 @@
 
 import { useCallback, useState } from 'react';
 
-/* eslint-disable react/no-multi-comp */
-
-import * as React from 'react';
+import ParameterControl from './ParameterControl';
 
 interface MapControlsProps {
   readonly padding: number;
@@ -22,63 +20,6 @@ interface MapControlsProps {
   readonly onRefreshLayout: () => void;
 }
 
-interface ParameterControlProps {
-  readonly explanation: string;
-  readonly id: string;
-  readonly label: string;
-  readonly max: number;
-  readonly min: number;
-  readonly onChange: (val: number) => void;
-  readonly step: number;
-  readonly value: number;
-}
-
-function ParameterControl({
-  label,
-  id,
-  value,
-  onChange,
-  min,
-  max,
-  step,
-  explanation = '',
-}: ParameterControlProps) {
-  const handleChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      onChange(parseFloat(e.target.value));
-    },
-    [onChange]
-  );
-
-  return (
-    <div className="map-control-group">
-      <label
-        className="map-control-label"
-        htmlFor={id}
-      >
-        {label}
-
-        :
-        {value.toFixed(step < 1 ? 2 : 0)}
-      </label>
-
-      <input
-        className="map-control-input"
-        id={id}
-        max={max}
-        min={min}
-        onChange={handleChange}
-        step={step}
-        type="range"
-        value={value}
-      />
-
-      {explanation ? <p className="map-control-explanation">
-        {explanation}
-      </p> : null}
-    </div>
-  );
-}
 
 /**
  * Collapsible panel for adjusting map layout parameters.
